requirements.txt
requests
beautifulsoup4
pandas
numpy
openpyxl
sqlalchemy
pymysql
sshtunnel
fastapi
uvicorn
mlxtend
python-dotenv